Refactor: improve structure, accessibility, and formatting
- Changed Gallery, GalleryResources, and scene classes from sealed to non-sealed for inheritance support.
- Moved scene classes to SkiaGallery.Scenes namespace.
- Made GalleryResources.CreateText static and updated all usages.
- Improved code formatting and consistency.
- Added missing using directives in App.cs and Gallery.cs.
- Added UTF-8 BOM to project and source files.
