+++ sources/Experiments/SkiaGallery/App.cs
@@ -1,4 +1,4 @@
-using System.Numerics;
+﻿using System.Numerics;
 using Silk.NET.Input;
 using Silk.NET.Windowing;
 using SkiaGallery.Helpers;
