+++ sources/Experiments/SkiaGallery/Helpers/CocoaHelper.cs
@@ -1,4 +1,4 @@
-using System.Runtime.InteropServices;
+﻿using System.Runtime.InteropServices;
 
 namespace SkiaGallery.Helpers;
