+++ sources/Experiments/SkiaGallery/Program.cs
@@ -1,3 +1,3 @@
-using SkiaGallery;
+﻿using SkiaGallery;
 
 App.Run();
